# docs/openapi.yml
openapi: 3.0.3
info:
  title: Flea Market System API
  version: 1.0.0
  description: |
    API definition for the Flea Market Matching Application.
    Provides endpoints for users (buyers/sellers) and administrators.
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
paths:
...
  /my-page:
    get:
      tags: [MyPage]
      summary: Get my profile
      description: Returns the authenticated user's profile information.
      security:
        - Bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    post:
      tags: [MyPage]
      summary: Update profile
      description: Update user name and profile image.
      security:
        - Bearer: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                profileImage:
                  type: string
                  format: binary
                  description: Profile image file
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  isSuccess:
                    type: boolean
                  errorMessage:
                    type: string
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List users
      security:
        - Bearer: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [id, name, email, banned]
            default: id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}:
    get:
      tags: [Admin]
      summary: Get user details
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    post:
      tags: [Admin]
      summary: Update user details
      description: Update user information (name, email, role, enabled status).
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [USER, ADMIN]
                enabled:
                  type: boolean
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, ADMIN]
        enabled:
          type: boolean
        banned:
          type: boolean
        banReason:
          type: string
        bannedAt:
          type: string
          format: date-time
        bannedByAdminId:
          type: integer
          format: int64
      required:
        - id
        - name
        - email
        - role
        - enabled
        - banned
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: API Key authentication via Bearer token